+++ sdk/azcore/arm/runtime/policy_bearer_token_test.go
@@ -5,9 +5,9 @@ package runtime
 
 import (
 	"context"
+	"errors"
 	"strings"
 
-	"errors"
 	"net/http"
 	"testing"
 	"time"
@@ -17,7 +17,9 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/shared"
 	azpolicy "github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/internal/errorinfo"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/mock"
+	"github.com/stretchr/testify/require"
 )
 
 const (
@@ -90,10 +92,10 @@ func TestBearerPolicy_SuccessGetToken(t *testing.T) {
 func TestBearerPolicy_CredentialFailGetToken(t *testing.T) {
 	srv, close := mock.NewTLSServer()
 	defer close()
-	expectedErr := errors.New("oops")
+	expectedErr := "oops"
 	failCredential := mockCredential{}
 	failCredential.getTokenImpl = func(ctx context.Context, options azpolicy.TokenRequestOptions) (azcore.AccessToken, error) {
-		return azcore.AccessToken{}, expectedErr
+		return azcore.AccessToken{}, errors.New(expectedErr)
 	}
 	b := NewBearerTokenPolicy(failCredential, nil)
 	pipeline := newTestPipeline(&azpolicy.ClientOptions{
@@ -104,16 +106,11 @@ func TestBearerPolicy_CredentialFailGetToken(t *testing.T) {
 		PerRetryPolicies: []azpolicy.Policy{b},
 	})
 	req, err := runtime.NewRequest(context.Background(), http.MethodGet, srv.URL())
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 	resp, err := pipeline.Do(req)
-	if err != expectedErr {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if resp != nil {
-		t.Fatal("expected nil response")
-	}
+	require.EqualError(t, err, expectedErr)
+	require.Nil(t, resp)
+	require.Implements(t, (*errorinfo.NonRetriable)(nil), err)
 }
 
 func TestBearerTokenPolicy_TokenExpired(t *testing.T) {
@@ -165,41 +162,42 @@ func TestBearerPolicy_GetTokenFailsNoDeadlock(t *testing.T) {
 	}
 }
 
-func TestBearerTokenWithAuxiliaryTenants(t *testing.T) {
-	t.Skip("this feature isn't implemented yet")
+func TestAuxiliaryTenants(t *testing.T) {
 	srv, close := mock.NewTLSServer()
 	defer close()
-	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
-	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
-	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
-	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
-	srv.AppendResponse()
-	retryOpts := azpolicy.RetryOptions{
-		MaxRetryDelay: 500 * time.Millisecond,
-		RetryDelay:    50 * time.Millisecond,
-	}
+	srv.SetResponse(mock.WithStatusCode(http.StatusOK))
+	primary := "primary"
+	auxTenants := []string{"aux1", "aux2", "aux3"}
+	expectCache := false
 	b := NewBearerTokenPolicy(
-		mockCredential{},
-		&armpolicy.BearerTokenOptions{
-			Scopes: []string{scope},
-			//AuxiliaryTenants: []string{"tenant1", "tenant2", "tenant3"},
+		mockCredential{
+			// getTokenImpl returns a token whose value equals the requested tenant so the test can validate how the policy handles tenants
+			// i.e., primary tenant token goes in Authorization header and aux tenant tokens go in x-ms-authorization-auxiliary
+			getTokenImpl: func(ctx context.Context, options azpolicy.TokenRequestOptions) (azcore.AccessToken, error) {
+				require.False(t, expectCache, "client should have used a cached token instead of requesting another")
+				tenant := primary
+				if options.TenantID != "" {
+					tenant = options.TenantID
+				}
+				return azcore.AccessToken{Token: tenant, ExpiresOn: time.Now().Add(time.Hour).UTC()}, nil
+			},
 		},
+		&armpolicy.BearerTokenOptions{AuxiliaryTenants: auxTenants, Scopes: []string{scope}},
 	)
-	pipeline := newTestPipeline(&azpolicy.ClientOptions{Transport: srv, Retry: retryOpts, PerRetryPolicies: []azpolicy.Policy{b}})
-	req, err := runtime.NewRequest(context.Background(), http.MethodGet, srv.URL())
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	resp, err := pipeline.Do(req)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("unexpected status code: %d", resp.StatusCode)
-	}
-	expectedHeader := strings.Repeat(shared.BearerTokenPrefix+tokenValue+", ", 3)
-	expectedHeader = expectedHeader[:len(expectedHeader)-2]
-	if auxH := resp.Request.Header.Get(shared.HeaderAuxiliaryAuthorization); auxH != expectedHeader {
-		t.Fatalf("unexpected auxiliary authorization header %s", auxH)
+	pipeline := newTestPipeline(&azpolicy.ClientOptions{Transport: srv, PerRetryPolicies: []azpolicy.Policy{b}})
+	expected := strings.Split(shared.BearerTokenPrefix+strings.Join(auxTenants, ","+shared.BearerTokenPrefix), ",")
+	for i := 0; i < 3; i++ {
+		if i == 1 {
+			// policy should have a cached token after the first iteration
+			expectCache = true
+		}
+		req, err := runtime.NewRequest(context.Background(), http.MethodGet, srv.URL())
+		require.NoError(t, err)
+		resp, err := pipeline.Do(req)
+		require.NoError(t, err)
+		require.Equal(t, shared.BearerTokenPrefix+primary, resp.Request.Header.Get(shared.HeaderAuthorization), "Authorization header must contain primary tenant token")
+		actual := strings.Split(resp.Request.Header.Get(headerAuxiliaryAuthorization), ", ")
+		// auxiliary tokens may appear in arbitrary order
+		require.ElementsMatch(t, expected, actual)
 	}
 }